Video: GWindow now correctly sized for application
Initially, when an SDL application launced, the Window size was
incredibly small. This has been rectified by making a call to
`GWindow::resize()` when the window is first constructed. The
SDL window already has the size passed from the application, hence
the window is set to the correct size.

A small list of 'default' video modes has been added too that are
added to the list that can be queried by the SDL application.
Realistically, these should come from `GWindow` itself.