Audio: Add a SerenityOS audio backend using /dev/audio
Ideally we would be using the AudioServer, but when trying to implement
this I ran into trouble with SDL clients trying to connect to the audio
IPC server from a secondary thread. This is currently not well handled
by the AClientConnection implementation in LibAudio.

The audio backend forces the audio spec to use 44100/16-bit/2-channel
samples since that's what the Serenity /dev/audio is expecting.